refactor: remove walletconnect
