+++ .env.example
@@ -1,4 +1,3 @@
 POSTHOG_API_HOST="Go to the PostHog project dashboard: Settings > Project Region"
 POSTHOG_PROJECT_ID="Go to the PostHog project dashboard: Settings > Project ID"
 PROVIDER_ID="Go to the GitHub repository: Settings > Secrets and variables > Actions > Variables > Repository variables"
-WALLET_CONNECT_PROJECT_ID="Go to WalletConnect and get project ID"
