+++ dapp-example/@types/global/index.d.ts
@@ -1,2 +1,1 @@
 declare const __PROVIDER_ID__: string;
-declare const __WALLET_CONNECT_PROJECT_ID__: string;
