+++ package.json
@@ -75,8 +75,6 @@
     "@types/zxcvbn": "^4.4.1",
     "@typescript-eslint/eslint-plugin": "^5.54.0",
     "@typescript-eslint/parser": "^5.54.0",
-    "@walletconnect/types": "^2.8.0",
-    "@web3modal/sign-react": "^2.4.5",
     "add": "^2.0.6",
     "addons-linter": "^6.17.0",
     "concurrently": "^7.6.0",
@@ -126,9 +124,6 @@
     "@stablelib/base64": "^1.0.1",
     "@stablelib/hex": "^1.0.1",
     "@stablelib/utf8": "^1.0.1",
-    "@walletconnect/core": "^2.8.0",
-    "@walletconnect/utils": "^2.8.0",
-    "@walletconnect/web3wallet": "^1.8.0",
     "algosdk": "^2.7.0",
     "bignumber.js": "^9.1.1",
     "buffer": "^6.0.3",
