+++ src/@types/global/index.d.ts
@@ -5,4 +5,3 @@ declare const __POSTHOG_PROJECT_ID__: string;
 declare const __PROVIDER_ID__: string;
 declare const __TARGET__: 'chrome' | 'edge' | 'firefox' | 'opera' | 'safari';
 declare const __VERSION__: string;
-declare const __WALLET_CONNECT_PROJECT_ID__: string;
