+++ src/extension/components/SessionAvatar/types/IProps.ts
@@ -1,0 +1,7 @@
+interface IProps {
+  iconUrl?: string;
+  name: string;
+  size?: 'sm' | 'md';
+}
+
+export default IProps;
