+++ src/extension/components/SessionAvatar/types/index.ts
@@ -1,0 +1,1 @@
+export type { default as IProps } from './IProps';
