--- src/extension/components/WalletConnectBannerIcon/index.ts
@@ -1,1 +1,0 @@
-export { default } from './WalletConnectBannerIcon';
