--- src/extension/components/WalletConnectIcon/index.ts
@@ -1,1 +1,0 @@
-export { default } from './WalletConnectIcon';
