+++ src/extension/features/sessions/thunks/index.ts
@@ -1,7 +1,5 @@
 export { default as clearSessionsThunk } from './clearSessionsThunk';
 export { default as fetchSessionsThunk } from './fetchSessionsThunk';
-export { default as initializeWalletConnectThunk } from './initializeWalletConnectThunk';
 export { default as removeAuthorizedAddressThunk } from './removeAuthorizedAddressThunk';
 export { default as removeSessionByIdThunk } from './removeSessionByIdThunk';
-export { default as removeSessionByTopicThunk } from './removeSessionByTopicThunk';
 export { default as setSessionThunk } from './setSessionThunk';
