+++ src/extension/features/sessions/utils/getInitialState.ts
@@ -1,13 +1,10 @@
 // types
-import { IState } from '../types';
+import type { IState } from '../types';
 
 export default function getInitialState(): IState {
   return {
     fetching: false,
-    initializingWalletConnect: false,
     items: [],
     saving: false,
-    web3Wallet: null,
-    walletConnectModalOpen: false,
   };
 }
