--- src/extension/hooks/useWalletConnect/index.ts
@@ -1,3 +1,0 @@
-export { default } from './useWalletConnect';
-export * from './types';
-export * from './utils';
