--- src/extension/hooks/useWalletConnect/types/index.ts
@@ -1,1 +1,0 @@
-export type { default as IUseWalletConnectState } from './IUseWalletConnectState';
