--- src/extension/hooks/useWalletConnect/utils/index.ts
@@ -1,1 +1,0 @@
-export { default as createSessionNamespaces } from './createSessionNamespaces';
