--- src/extension/modals/WalletConnectModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './WalletConnectModal';
