+++ src/extension/selectors/index.ts
@@ -17,4 +17,3 @@ export * from './sessions';
 export * from './settings';
 export * from './standard-assets';
 export * from './system';
-export * from './walletconnect';
