--- src/extension/selectors/walletconnect/index.ts
@@ -1,3 +1,0 @@
-export { default as useSelectInitializingWalletConnect } from './useSelectInitializingWalletConnect';
-export { default as useSelectWalletConnectModalOpen } from './useSelectWalletConnectModalOpen';
-export { default as useSelectWeb3Wallet } from './useSelectWeb3Wallet';
