+++ src/extension/types/sessions/ISession.ts
@@ -1,6 +1,3 @@
-// types
-import type IWalletConnectSessionMetadata from '../IWalletConnectSessionMetadata';
-
 /**
  * @property {string} appName - the name of the dApp.
  * @property {string[]} authorizedAddresses - a list of authorized addresses.
@@ -12,7 +9,6 @@ import type IWalletConnectSessionMetadata from '../IWalletConnectSessionMetadata
  * @property {string | null} iconUrl - an icon url for the app.
  * @property {string} id - a unique identifier for this session as a UUID v4.
  * @property {number} usedAt - a timestamp (in milliseconds) for when this session was last used.
- * @property {IWalletConnectSessionMetadata | null} walletConnectMetadata - extra data if this is a WalletConnect session.
  */
 interface ISession {
   appName: string;
@@ -25,7 +21,6 @@ interface ISession {
   iconUrl: string | null;
   id: string;
   usedAt: number;
-  walletConnectMetadata: IWalletConnectSessionMetadata | null;
 }
 
 export default ISession;
