+++ src/extension/utils/mapSessionFromEnableRequest/index.ts
@@ -1,1 +1,2 @@
 export { default } from './mapSessionFromEnableRequest';
+export * from './types';
