+++ src/extension/utils/mapSessionFromEnableRequest/types/IOptions.ts
@@ -1,0 +1,11 @@
+// types
+import type { IClientInformation } from '@common/types';
+import type { INetwork } from '@extension/types';
+
+interface IOptions {
+  authorizedAddresses: string[];
+  clientInfo: IClientInformation;
+  network: INetwork;
+}
+
+export default IOptions;
