+++ src/extension/utils/mapSessionFromEnableRequest/types/index.ts
@@ -1,0 +1,1 @@
+export type { default as IOptions } from './IOptions';
