--- src/extension/utils/mapSessionFromWalletConnectSession/index.ts
@@ -1,1 +1,0 @@
-export { default } from './mapSessionFromWalletConnectSession';
