+++ test/setup.ts
@@ -1,6 +1,5 @@
 import 'isomorphic-fetch';
 import { TextDecoder, TextEncoder } from 'util';
-import { v4 as uuid } from 'uuid';
 
 // config
 import { name, version } from '../package.json';
@@ -16,9 +15,6 @@ Object.defineProperty(global, '__TARGET__', {
 });
 Object.defineProperty(global, '__VERSION__', {
   value: `${version}-test`,
-});
-Object.defineProperty(global, '__WALLET_CONNECT_PROJECT_ID__', {
-  value: uuid(),
 });
 Object.defineProperty(global, 'TextDecoder', {
   value: TextDecoder,
